pyproject.toml
[project]
name = "artemis-cve"
version = "0.1.0"
description = ""
authors = [
    {name = "wangchenyu",email = "1542169671@qq.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "grpcio (>=1.78.0,<2.0.0)",
    "protobuf (>=6.31.1,<7.0.0)",
    "numpy (==1.26.4)",
    "opencv-python (<4.13)",
    "pillow (>=12.1.1,<13.0.0)",
    "tqdm (>=4.67.3,<5.0.0)",
    "scipy (<1.16)",
    "typer (>=0.24.1,<0.25.0)",
    "aiortc (>=1.14.0,<2.0.0)",
    "safetensors (>=0.7.0,<0.8.0)",
    "transformers (>=5.3.0,<6.0.0)",
    "torch (>=2.4.0,<3.0.0)",
    "ultralytics (>=8.3.0,<9.0.0)"
]

[tool.poetry]
packages = [{include = "artemis_cve", from = "src"}]

[tool.poetry.group.dev.dependencies]
grpcio-tools = ">=1.78.0,<2.0.0"
mypy-protobuf = ">=3.6.0,<4.0.0"
types-protobuf = ">=6.30.2,<7.0.0"

[tool.poetry.scripts]
artemis-cve = "artemis_cve.commands.app:app"

[[tool.poetry.source]]
name = "tuna"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
priority = "primary"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
